+++ config.cmake.in
@@ -16,6 +16,7 @@ if(@PCMS_ENABLE_OMEGA_H@)
     find_dependency(Omega_h CONFIG HINTS @Omega_h_DIR@)
 endif()
 
+include("${CMAKE_CURRENT_LIST_DIR}/pcms_utility-targets.cmake")
 include("${CMAKE_CURRENT_LIST_DIR}/pcms_core-targets.cmake")
 include("${CMAKE_CURRENT_LIST_DIR}/pcms_interpolator-targets.cmake")
