+++ examples/external-usage-example/main.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-#include <pcms/print.h>
+#include <pcms/utility/print.h>
 
 int main()
 {
