+++ src/CMakeLists.txt
@@ -2,10 +2,7 @@
 set(
   PCMS_HEADERS
   pcms.h
-  pcms/arrays.h
-  pcms/assert.h
   pcms/bounding_box.h
-  pcms/common.h
   pcms/coordinate.h
   pcms/coordinate_systems.h
   pcms/coordinate_transform.h
@@ -14,12 +11,6 @@ set(
   pcms/field_communicator.h
   pcms/field_communicator2.h
   pcms/field_evaluation_methods.h
-  pcms/memory_spaces.h
-  pcms/types.h
-  pcms/array_mask.h
-  pcms/inclusive_scan.h
-  pcms/profile.h
-  pcms/print.h
   pcms/partition.h
   pcms/coupler.h
   pcms/coordinate_system.h
@@ -33,8 +24,6 @@ set(
 set(
   PCMS_SOURCES
   pcms.cpp
-  pcms/assert.cpp
-  pcms/print.cpp
   pcms/create_field.cpp
   pcms/adapter/xgc/xgc_field_adapter.h
   pcms/adapter/point_cloud/point_cloud_layout.cpp
@@ -47,6 +36,8 @@ set(
 configure_file(pcms/version.h.in pcms/version.h)
 configure_file(pcms/configuration.h.in pcms/configuration.h)
 list(APPEND PCMS_HEADERS ${CMAKE_CURRENT_BINARY_DIR}/pcms/version.h ${CMAKE_CURRENT_BINARY_DIR}/pcms/configuration.h)
+
+add_subdirectory(pcms/utility)
 
 
 if(PCMS_ENABLE_XGC)
@@ -78,7 +69,7 @@ add_library(pcms::core ALIAS pcms_core)
 target_compile_features(pcms_core PUBLIC cxx_std_17)
 target_link_libraries(
   pcms_core PUBLIC meshfields::meshfields redev::redev
-  MPI::MPI_CXX Kokkos::kokkos perfstubs
+  MPI::MPI_CXX Kokkos::kokkos perfstubs pcms::utility
 )
 if(PCMS_ENABLE_OMEGA_H)
   target_link_libraries(pcms_core PUBLIC Omega_h::omega_h)
@@ -89,10 +80,6 @@ if(PCMS_HAS_ASAN)
     pcms_core PRIVATE -fsanitize=address
     -fno-omit-frame-pointer
   )
-endif()
-
-if(PCMS_ENABLE_SPDLOG)
-  target_link_libraries(pcms_core PUBLIC spdlog::spdlog)
 endif()
 
 # export the library
