+++ src/pcms.h
@@ -2,10 +2,10 @@
 #define PCMS_H_
 
 #include "pcms/configuration.h"
-#include "pcms/common.h"
+#include "pcms/utility/common.h"
+#include "pcms/utility/profile.h"
 #include "pcms/field_communicator.h"
 #include "pcms/adapter/omega_h/omega_h_field.h"
-#include "pcms/profile.h"
 #include "pcms/coupler.h"
 
 #endif
