+++ src/pcms/adapter/xgc/xgc_reverse_classification.cpp
@@ -1,7 +1,7 @@
 #include "xgc_reverse_classification.h"
 #include "mpi.h"
 #include <fstream>
-#include "pcms/assert.h"
+#include "pcms/utility/assert.h"
 #include <string>
 namespace pcms
 {
