+++ src/pcms/coordinate.h
@@ -1,7 +1,7 @@
 #ifndef PCMS_COUPLING_COORDINATE_H
 #define PCMS_COUPLING_COORDINATE_H
 #include <array>
-#include "pcms/types.h"
+#include "pcms/utility/types.h"
 #include <Kokkos_Macros.hpp>
 namespace pcms
 {
