+++ src/pcms/field.h
@@ -1,9 +1,9 @@
 #ifndef PCMS_COUPLING_FIELD_H
 #define PCMS_COUPLING_FIELD_H
 #include "field_layout.h"
-#include "pcms/types.h"
-#include "pcms/arrays.h"
-#include "pcms/memory_spaces.h"
+#include "pcms/utility/types.h"
+#include "pcms/utility/arrays.h"
+#include "pcms/utility/memory_spaces.h"
 #include <map>
 #include <memory>
 #include <string>
