+++ src/pcms/interpolator/pcms_interpolator_aliases.hpp
@@ -2,7 +2,7 @@
 #define PCMS_INTERPOLATOR_ALIASES_HPP
 
 #include <Kokkos_Core.hpp>
-#include "pcms/arrays.h"
+#include "pcms/utility/arrays.h"
 
 namespace pcms
 {
