+++ src/pcms/utility/assert.cpp
@@ -1,7 +1,5 @@
-#include "pcms/assert.h"
-#include "pcms/print.h"
-#include <cstdio>
-#include <cstdlib>
+#include "pcms/utility/print.h"
+
 namespace pcms
 {
 void Pcms_Assert_Fail(const char* msg)
