+++ src/pcms/utility/assert.h
@@ -1,18 +1,19 @@
 #ifndef PCMS_COUPLING_ASSERT_H
 #define PCMS_COUPLING_ASSERT_H
-#include <cassert>
 #include <mpi.h>
 
 // https://stackoverflow.com/questions/16683146/can-macros-be-overloaded-by-number-of-arguments
-#define CAT(A, B) A##B
-#define SELECT(NAME, NUM) CAT(NAME##_, NUM)
+#define PCMS_CAT(A, B) A##B
+#define PCMS_SELECT(NAME, NUM) PCMS_CAT(NAME##_, NUM)
 
-#define GET_COUNT(_1, _2, _3, _4, _5, _6 /* ad nauseam */, COUNT, ...) COUNT
-#define VA_SIZE(...) GET_COUNT(__VA_ARGS__, 6, 5, 4, 3, 2, 1)
+#define PCMS_GET_COUNT(_1, _2, _3, _4, _5, _6 /* ad nauseam */, COUNT, ...)    \
+  COUNT
+#define PCMS_VA_SIZE(...) PCMS_GET_COUNT(__VA_ARGS__, 6, 5, 4, 3, 2, 1)
 
-#define VA_SELECT(NAME, ...) SELECT(NAME, VA_SIZE(__VA_ARGS__))(__VA_ARGS__)
+#define PCMS_VA_SELECT(NAME, ...)                                              \
+  PCMS_SELECT(NAME, PCMS_VA_SIZE(__VA_ARGS__))(__VA_ARGS__)
 
-#define PCMS_ALWAYS_ASSERT(...) VA_SELECT(PCMS_ALWAYS_ASSERT, __VA_ARGS__)
+#define PCMS_ALWAYS_ASSERT(...) PCMS_VA_SELECT(PCMS_ALWAYS_ASSERT, __VA_ARGS__)
 #define PCMS_ALWAYS_ASSERT_1(cond) PCMS_ALWAYS_ASSERT_2(cond, MPI_COMM_WORLD)
 #define PCMS_ALWAYS_ASSERT_2(pcms_macro_cond, pcms_macro_comm)                                                 \
   /* NOLINTBEGIN(modernize-avoid-c-arrays,cppcoreguidelines-avoid-do-while,cppcoreguidelines-avoid-c-arrays)*/ \
