+++ src/pcms/utility/print.cpp
@@ -1,4 +1,7 @@
-#include "print.h"
+#include "pcms/utility/print.h"
+
+#include <cstdio>
+#include <cassert>
 
 namespace pcms
 {
