+++ test/xgc_n0_coupling_server.cpp
@@ -1,5 +1,5 @@
 #include <pcms.h>
-#include <pcms/types.h>
+#include <pcms/utility/types.h>
 #include <Omega_h_file.hpp>
 #include <Omega_h_for.hpp>
 #include "test_support.h"
